+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.29
+
+* Unfilling of `UIKitText`
+
 ## 0.0.28
 
 * Fixes in `UIKitWidth`